LayoutTests/platform/win/fast/forms/reset-autofilled-expected.txt

This tests that a field's autofilled state is reverted when the form is reset. It can only be run using DumpRenderTree.

PASS countAutofilledElements() is 1
FAIL countAutofilledElements() should be 0. Was 1.
